Permit a broad range of Husky versions
After merging #2 I realized the Husky version was caret-ranged, and
since "stable" versions of Husky are still in pre-release (i.e. the
major version number is 0) this locks consumers at minor Husky
versions, which is likely to cause some inconvenience and makes the
switch to ^0.14.3 a breaking change.

I'd like to make the next release compatible with Husky 1.x based on
release candidates, so I'm replacing 1.0's ^0.13.3 with a primitive
range to reflect that intent.